+++ src/main/java/roomescape/reservationtime/controller/ReservationTimeController.java
@@ -1,6 +1,8 @@
 package roomescape.reservationtime.controller;
 
 import jakarta.validation.Valid;
+import jakarta.validation.constraints.NotNull;
+import java.time.LocalDate;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
@@ -10,8 +12,10 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
+import roomescape.reservationtime.dto.AvailableReservationTimeResponse;
 import roomescape.reservationtime.dto.ReservationTimeRequest;
 import roomescape.reservationtime.dto.ReservationTimeResponse;
 import roomescape.reservationtime.service.ReservationTimeService;
@@ -27,6 +31,14 @@ public class ReservationTimeController {
     @ResponseStatus(HttpStatus.CREATED)
     public ReservationTimeResponse saveTime(@Valid @RequestBody final ReservationTimeRequest request) {
         return reservationTimeService.saveTime(request);
+    }
+
+    @GetMapping("/reservations")
+    public List<AvailableReservationTimeResponse> findAllAvailableTimes(
+            @NotNull @RequestParam final LocalDate date,
+            @NotNull @RequestParam final Long themeId
+    ) {
+        return reservationTimeService.findAllReservationTime(date, themeId);
     }
 
     @GetMapping
