+++ src/test/java/roomescape/reservation/service/WaitingReservationServiceTest.java
@@ -137,7 +137,7 @@ class WaitingReservationServiceTest {
         waitingReservationRepository.save(waitingReservation);
 
         // when
-        service.deleteById(waitingReservation.getId());
+        service.cancel(waitingReservation.getId());
 
         // then
         assertThat(waitingReservationRepository.findById(waitingReservation.getId())).isEmpty();
